refactor: move mock data and types to separate files
